+++ IBSWeb/Areas/SuperAdmin/Controllers/HomeController.cs
@@ -1,0 +1,15 @@
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+namespace IBSWeb.Areas.SuperAdmin.Controllers
+{
+    [Area("SuperAdmin")]
+    [Authorize(Roles = "SuperAdmin")]
+    public class HomeController : Controller
+    {
+        public IActionResult Index()
+        {
+            return View();
+        }
+    }
+}